requirements.txt
streamlit
pandas
numpy
openpyxl
google-generativeai
